requirements.txt
accelerate==0.23.0
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.6
decorator==4.4.2
diffusers==0.21.4
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.12.4
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
Jinja2==3.1.2
lit==17.0.2
MarkupSafe==2.1.3
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.2
pathtools==0.1.2
Pillow==10.0.1
proglog==0.1.10
protobuf==4.24.3
psutil==5.9.5
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.3.3
sentry-sdk==1.31.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tokenizers==0.14.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.34.0
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.6
wandb==0.15.11
zipp==3.17.0
requests
beautifulsoup4
urllib3
omegaconf
cloudscraper